# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "寵知 API",
    "version": "1.0.0",
    "description": "API 文件"
  },
  "host": "petknow-be.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "authorization",
      "description": "請加上 API Token"
    }
  },
  "paths": {
    "/v1/register": {
      "post": {
        "tags": [
          "登入系統 API"
        ],
        "description": "註冊帳號",
        "parameters": [
          {
            "name": "body",
            "description": "資料格式",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Benson"
                },
                "email": {
                  "type": "string",
                  "example": "Abc1231@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "Abc123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "statusCode": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Benson"
                    },
                    "email": {
                      "type": "string",
                      "example": "testfge2@gmail.com"
                    },
                    "__v": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "success": false,
            "statusCode": 400,
            "message": "Failure",
            "description": "Bad Request"
          },
          "500": {
            "description": "伺服器發生錯誤",
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "number",
                  "example": 500
                },
                "isSuccess": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "系統發生錯誤，請聯繫系統管理員"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/v1/login": {
      "post": {
        "tags": [
          "登入系統 API"
        ],
        "description": "註冊帳號",
        "parameters": [
          {
            "name": "body",
            "description": "資料格式",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "Abc1231@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "Abc123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "statusCode": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Benson"
                    },
                    "email": {
                      "type": "string",
                      "example": "Abc1231@gmail.com"
                    },
                    "__v": {
                      "type": "number",
                      "example": 0
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDdmZDVlZTVkNWMxYTNjNTZlYTk0MTQiLCJpYXQiOjE2ODYyODAxMjZ9.t2RGvEU_AICIA38mAekXENBEWWKQl54jeOEztgruZtg"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "錯誤的請求",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "statusCode": {
                  "type": "number",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Invalid credentials"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "Abc1231@gmail.com"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "伺服器發生錯誤",
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "number",
                  "example": 500
                },
                "isSuccess": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "系統發生錯誤，請聯繫系統管理員"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/v1/user/show": {
      "get": {
        "tags": [
          "登入系統 API"
        ],
        "description": "註冊帳號",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "statusCode": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "646d8bf85f52ae9681b88593"
                    },
                    "name": {
                      "type": "string",
                      "example": "Benson"
                    },
                    "email": {
                      "type": "string",
                      "example": "Abc123@gmail.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$10$1yR.UPzTukYyVdiLUTHvhugqp1nBEjZmKD31inX2XGlLpUIGvAwly"
                    },
                    "__v": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "錯誤的請求",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "statusCode": {
                  "type": "number",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "未通過認證 或 查詢不到使用者資料"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "伺服器發生錯誤",
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "number",
                  "example": 500
                },
                "isSuccess": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "系統發生錯誤，請聯繫系統管理員"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/v1/user/update": {
      "put": {
        "tags": [
          "登入系統 API"
        ],
        "description": "註冊帳號",
        "parameters": [
          {
            "name": "body",
            "description": "資料格式",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "bio": {
                  "type": "string",
                  "example": "無敵鐵金剛"
                },
                "nickname": {
                  "type": "string",
                  "example": "說你好"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "number",
                  "example": 200
                },
                "isSuccess": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "更新成功"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "patrick"
                    },
                    "email": {
                      "type": "string",
                      "example": "Abc1231@gmail.com"
                    },
                    "__v": {
                      "type": "number",
                      "example": 0
                    },
                    "bio": {
                      "type": "string",
                      "example": "無敵鐵金剛"
                    },
                    "nickname": {
                      "type": "string",
                      "example": "說你好"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "錯誤的請求",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "statusCode": {
                  "type": "number",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "未通過認證 或 查詢不到使用者資料"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "伺服器發生錯誤",
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "number",
                  "example": 500
                },
                "isSuccess": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "系統發生錯誤，請聯繫系統管理員"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  }
}